Expose root variant in ResolutionResult
There is no way to get the root variant of the root component from a ResolutionResult without looping through all varaints and finding the root variant with its display name.

This commit exposes the root varaint so that users do not need to rely on variant display names to begin traversing dependency graphs.