Generate Override annotation and package name for precompiled script plugin adapters
For precompiled Groovy script plugins, we generate an adapter class. This class is written in Java, and is the
actual plugin that Gradle interacts with. Then, the adapter executes the precompiled script.

We generate this adapter as source code. Newer compilers complain that the generated source code is missing
an @Override annotation on its apply method. They also complain that the generated class lives in the default package.

This commit updates the generated code to add the missing annotation and to live in the 'precompiled' package.

For example, if the old generated class name was FooPlugin, the new class name is precompiled.FooPlugin